+++ user-operator/src/test/java/io/strimzi/operator/user/model/KafkaUserModelCertificateHandlingTest.java
@@ -6,9 +6,9 @@ package io.strimzi.operator.user.model;
 
 import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.api.model.SecretBuilder;
-import io.strimzi.api.ResourceAnnotations;
 import io.strimzi.certs.CertAndKey;
 import io.strimzi.certs.CertManager;
+import io.strimzi.operator.common.Annotations;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.model.Ca;
 import io.strimzi.operator.common.model.Labels;
@@ -75,7 +75,7 @@ public class KafkaUserModelCertificateHandlingTest {
                 .withNewMetadata()
                     .withName(ResourceUtils.NAME)
                     .withNamespace(ResourceUtils.NAMESPACE)
-                    .addToAnnotations(ResourceAnnotations.ANNO_STRIMZI_IO_FORCE_RENEW, "true")
+                    .addToAnnotations(Annotations.ANNO_STRIMZI_IO_FORCE_RENEW, "true")
                 .endMetadata()
                 .build();