+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Enable configuring `allowList` of Strimzi Metrics Reporter dynamically
 * Add support for TLS/SSL on the HTTP Bridge
   Set `spec.http.tls.certificateAndKey` configuration to enable it and provide the certificate and key via Secret.
+* Add support force-renewal of KafkaUser certificates via `strimzi.io/force-renew` annotation
 
 ### Major changes, deprecations, and removals
 
@@ -30,7 +31,6 @@
 * Update HTTP bridge to 1.0.0.
   * `/metrics` endpoint is no longer available on the regular HTTP interface (port 8080 by default). It is now available on the HTTP management interface, 8081.
     Users upgrading to Strimzi 1.0.0+ should check all monitoring configurations that scrape Kafka Bridge metrics and update them to use port 8081 instead of 8080 or any other non-default port before or immediately after the upgrade to avoid metrics collection failures.
-* Support force-renewal of KafkaUser certificates via `strimzi.io/force-renew` annotation
 
 ## 0.51.0